improved Converter namespace
